move emitter implementation to a trait
